// src/interfaces/RegisterHoursInterface.tsx
// Generated by https://quicktype.io

export interface getCollectionParams {
  rowsPerPage: number,
  fechaProgInicial?: string,
  fechaProgFinal?: string,
  reload?: boolean,
}

export interface updateParams {
  id: number,
  fecha_prestacion_servicio: string,
  hora_inicio_servicio: string,
  hora_final_servicio: string,
  observaciones_ejecucion: string,
  tipo_proceso: string,
  accion: string
}

export interface RegisterHours {
  datos:         Dato[];
  desde:         number;
  hasta:         number;
  por_pagina:    number;
  pagina_actual: number;
  ultima_pagina: number;
  total:         number;
}

export interface Dato {
  id:                         number;
  numero_orden_servicio:      number;
  fecha_creacion:             string;
  estado_orden_servicio:      EstadoOrdenServicio;
  asociado:                   string;
  fecha_programada:           string;
  hora_programada:            string;
  fecha_prestacion_servicio:  null | string;
  hora_inicio_servicio:       null | string;
  hora_final_servicio:        null | string;
  observaciones_ejecucion:    null | string;
  departamento:               string;
  departamento_id:            number;
  ciudad:                     string;
  lugar:                      string;
  direccion:                  string;
  recurso:                    null | string;
  tipo_servicio:              TipoServicio;
  numero_viaje:               number | null;
  recurso_id:                 number | null;
  equipo_id:                  number;
  numero_serial:              string;
  equipo:                     Equipo;
  observaciones_programacion: null | string;
  indicativo_aceptacion:      null | string;
  observaciones_rechazo:      null | string;
}

export enum Equipo {
  Detc = "DETC",
  Integracion02112 = "Integracion 0211-2",
  RotorGps = "Rotor-Gps",
}

export enum EstadoOrdenServicio {
  Rut = "RUT",
  Ter = "TER",
}

export enum TipoServicio {
  Desinstalación = "Desinstalación",
  Instalación = "Instalación",
}